Auto-reconnect to server
Also made the logs chattier to try and figure out what is going on.

I added "reliable" connections but I'm not sure if that matters - it
still drops the connections inside my wifi router, but at least the
second try regains the call!

I think the system is using UDP in the background, so we need to
add some periodic traffic to keep the NAT routes warm. This is true for
the peer connections, but possibly _also_ for the connection to the
server.