Allow nodes with SLOW clocks to still participate in the network
If they think a block is from the future, reevaluate its state the next
time some other node builds on that "future block". That way we can
still choose _close_ to the right blocks instead of hard-forking when
our clock is behind everybody else's.

If we're slower than the "main strength" of the network, we will waste
time hashing a block or two behind everyone else, but still identify
the best chain eventually.

If we're faster than the majority, then they will ignore our blocks
and we will fall behind on hashing power.