Tweaking the "fetching" algorithm a little bit more
It can now somewhat figure out which blocks are already known
by the other side, based on the overlaps between the two chains.

It still needs some work, since it's only good if the other side
knows about the block that you claim to "have" (which is NOT true
if you have been disconnected for a while and made your own "latest"
block).
Maybe we should send a list of "have" blocks, representing
the N most recent blocks on our "good chain" or something?
That way the other side can have a better chance of figuring out the overlap.