Limit incoming connection count to 4
This is to prevent individual clients from getting overwhelmed
with traffic. It also cuts down on the number of connections
that could be made for a larger network feeding through
the TURN server.

The number of connections is (Num nodes) x (Connections per node),
which can get astronomically huge for a fully-connected mesh! So
we need some sort of sanity limit.

Also, it's a cooler illustration of a peer-to-peer gossip network if
it is a more loosely-connected mesh.